Split hw into two files
