+++ hws/shunting_yard/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_executable(shunting_yard main.c)
+add_executable(shunting_yard main.c yard.c)
 target_link_libraries(shunting_yard PRIVATE stdhw)
